Remove references to x86 package no longer built
The …-i386.msi installer packages are no longer being included for new
releases, so all references to 32-bit installs were removed from the
packaging files. Also, library conflicts required a switch to a
versioned install TargetDir, so the cmdlet that generates the shim for
`qalc.exe` now targets that directory.

The Description in the `.nuspec` file was rewritten taking advantage of
the Markdown syntax parsing, based on the project website's Features
page, and likewise with the ReleaseNotes sourced on from the News page.

An explicit timeout value was set for the Update-Package function at 45
seconds, owing to a random local Internet outage highlighting the
fallback value of 100s as unhelpful.

Lastly, pursuant to the exchange in my last PR, I added myself to the
Owners tag as a co-maintainer. That was suggested to me as the path to
transferring maintainership with the least friction. The process can
then proceed as you think best.

Signed-off-by: Peter J. Mello <admin@petermello.net>