Add README.md file to qalculate package folder
To make the package directory more user-friendly (it's linked from the
package page on community.chocolatey.org), this commit adapts the
Markdown information added to the `.nuspec` file in commit c0d8b460 for
a README file GitHub will show automatically to visitors to the package
source dir. An optimized svg of the app icon was added for aesthetics
there, likewise with two screenshot PNG files.

Signed-off-by: Peter J. Mello <admin@petermello.net>